server/requirements.txt
Click==7.0
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
DateTime==4.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
plotly==4.5.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
retrying==1.3.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
zope.interface==4.7.1
